requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
psycopg[binary]
gunicorn
python-dotenv

